+++ R/summary_functions.R
@@ -20,7 +20,7 @@
 #'
 #' @examples
 #' # continuous variable summaries
-#' ard_continuous(
+#' ard_summary(
 #'   ADSL,
 #'   variables = "AGE",
 #'   statistic = ~ continuous_summary_fns(c("N", "median"))