+++ tests/testthat/test-print.R
@@ -1,19 +1,19 @@
 test_that("print.card() works", {
   expect_snapshot(
-    ard_continuous(ADSL, by = "ARM", variables = "AGE")
+    ard_summary(ADSL, by = "ARM", variables = "AGE")
   )
 
   expect_snapshot(
-    ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")
+    ard_tabulate(ADSL, by = "ARM", variables = "AGEGR1")
   )
 
   expect_snapshot(
-    ard_continuous(ADSL, variables = "AGE", fmt_fun = AGE ~ list(~ \(x) round(x, 3)))
+    ard_summary(ADSL, variables = "AGE", fmt_fun = AGE ~ list(~ \(x) round(x, 3)))
   )
 
   # checking the print of Dates
   expect_snapshot(
-    ard_continuous(
+    ard_summary(
       data = data.frame(x = seq(as.Date("2000-01-01"), length.out = 10L, by = "day")),
       variables = x,
       statistic = ~ continuous_summary_fns(c("min", "max", "sd"))
@@ -25,7 +25,7 @@ test_that("print.card() works", {
   expect_snapshot(
     bind_ard(
       ard_attributes(mtcars, variables = mpg),
-      ard_continuous(
+      ard_summary(
         mtcars,
         variables = mpg,
         statistic =