+++ tests/testthat/test-sort_ard_hierarchical.R
@@ -423,7 +423,7 @@ test_that("sort_ard_hierarchical() works with overall data", {
 test_that("sort_ard_hierarchical() error messaging works", {
   # invalid x input
   expect_snapshot(
-    sort_ard_hierarchical(ard_categorical(
+    sort_ard_hierarchical(ard_tabulate(
       ADSL,
       by = "ARM",
       variables = "AGEGR1"